agent_0

+++ contracts/agent/IAgentV2Immutables.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.20;
 
 interface IAgentV2Immutables {
-    function maxAgents() external view returns (uint256);
+    function maxAgents() external view returns (uint128);
 
     function factory() external view returns (address);
 }
